requirements.txt
langchain==0.1.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
# GitHub Copilot을 사용하기 위한 패키지
